fix: Clear the sign-compare warnings (#594)
Thirty-nine sites in fourteen files, nearly all a loop counter declared
`int` against a `size()`. Measure these with asserts enabled: a Release
build compiles away sixteen of them, including every one in
`term_translator.cpp` and the substitute tests, so the Release-only
count of 37 raw warnings is really 53.

One was a defect rather than hygiene.
`GenericSolver::get_assumptions_from_string` subtracted one from a
`find` result before asserting that result was not `npos`; because the
subtraction came first, a missing `)` left `end` one below `npos` and
the assert could never fire. Finding the position, checking it, and only
then subtracting gives the guard something to catch. The same shape
elsewhere is accidentally sound — an unadjusted `npos` stored in an
`int` is `-1`, which converts back to `npos` — but it only works while
nothing touches the value in between. `BoolectorTermIter::operator*` now
also says that `idx` must be non-negative, which is what comparing it to
a `size()` assumed.

+++ cvc5/src/cvc5_solver.cpp
@@ -900,7 +900,7 @@ Term Cvc5Solver::get_tester(const Sort & s, std::string name) const
   {
     std::shared_ptr<Cvc5Sort> cs = std::static_pointer_cast<Cvc5Sort>(s);
     cvc5::Datatype dt = cs->sort.getDatatype();
-    for (int i = 0; i != dt.getNumConstructors(); i++)
+    for (size_t i = 0; i != dt.getNumConstructors(); i++)
     {
       cvc5::DatatypeConstructor ct = dt[i];
       if (ct.getName() == name)
