+++ z3/include/z3_datatype.h
@@ -43,7 +43,7 @@ class Z3Datatype : public AbsDatatype
   }
   int get_num_selectors(std::string name) const override
   {
-    for (size_t i = 0; i < get_num_constructors(); i++)
+    for (int i = 0; i < get_num_constructors(); i++)
     {
       z3::func_decl cons{ c, Z3_get_datatype_sort_constructor(c, datatype, i) };
       if (cons.name().str() == name) return cons.arity();
